+++ backend/config/filesystems.php
@@ -53,6 +53,12 @@ return [
             'throw' => false,
         ],
 
+        'rem-discovery' => [
+            'driver' => 'local',
+            'root' => storage_path('app/rem-discovery'),
+            'throw' => false,
+        ],
+
         's3' => [
             'driver' => 's3',
             'key' => env('AWS_ACCESS_KEY_ID'),